Change generate random shop
The random shop will only generate if there is nothing in the item list.  It will also now never generate
a shop that contains nothing.

+++ GameSquire/AddItemViewController.swift
@@ -22,7 +22,7 @@ class AddItemViewController: UIViewController {
     }
     
     @IBAction func addBTN(_ sender: UIButton) {
-        if Int(priceTF.text!) != nil && Int(quantityTF.text!) != nil{
+        if Double(priceTF.text!) != nil && Int(quantityTF.text!) != nil{
             AppDelegate.myModel.addItem(name: itemTF.text!, price: Double(priceTF.text!)!, description: descriptionTF.text!, quantity: Int(quantityTF.text!)!)
         }
     }
